requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
boto3==1.29.3
botocore==1.32.6
certifi==2023.11.17
charset-normalizer==3.3.2
Deprecated==1.2.14
dune-client==1.3.1
frozendict==2.3.9
frozenlist==1.4.0
idna==3.6
jmespath==1.0.1
moralis==0.1.42
multidict==6.0.4
ndjson==0.3.1
numpy==1.26.2
pandas==2.1.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
types-Deprecated==1.2.9.3
types-python-dateutil==2.8.19.14
types-PyYAML==6.0.12.12
types-requests==2.31.0.6
types-setuptools==68.2.0.2
types-urllib3==1.26.25.14
typing_extensions==4.3.0
tzdata==2023.3
urllib3==1.26.18
wrapt==1.16.0
yarl==1.9.3
